+++ frontend/marketplace/src/graphql/queries/creditOffsetCertificate.ts
@@ -1,0 +1,17 @@
+import gql from "graphql-tag";
+
+export const GET_CREDIT_OFFSET_CERTIFICATE = gql`
+  query GetCreditOffsetCertificate($id: String!) {
+    creditOffsetCertificates(filter: { id: { equalTo: $id } }) {
+      nodes {
+        id
+        nodeId
+        denom
+        retiringEntityName
+        retiringEntityAdditionalData
+        walletId
+        amount
+      }
+    }
+  }
+`;
