+++ frontend/marketplace/src/graphql/queries/index.ts
@@ -1,2 +1,3 @@
 export * from "./creditCollections";
 export * from "./creditOffsetCertificates";
+export * from "./creditOffsetCertificate";
